Format code

+++ src/tomasulo/memory/MainMemory.java
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Stack;
 
 public class MainMemory {
+
 	public static int memoryCapacity = 64 * 1024;
 	private Block[] blocks;
 	private int blockSizeInBytes;
@@ -31,26 +32,27 @@ public class MainMemory {
 		Stack<String> instructionsStack = this.convertInstructionsArrayToStack(instructions);
 		while (!instructionsStack.isEmpty()) {
 			for (int i = 0; i < this.blockSizeInBytes / 4; i++) {
-				if (!instructionsStack.isEmpty()){
-					this.blocks[(int)startAddressInBytes / this.blockSizeInBytes].addData(instructionsStack.pop(), i);
+				if (!instructionsStack.isEmpty()) {
+					this.blocks[startAddressInBytes / this.blockSizeInBytes].addData(instructionsStack.pop(), i);
 				}
 			}
-			
+
 			startAddressInBytes += this.blockSizeInBytes;
 		}
 	}
-	
+
 	public Block readBlock(int addressInBytes) {
-		return this.blocks[(int)addressInBytes/4];
+		return this.blocks[addressInBytes / 4];
 	}
 
 	public static void main(String[] args) {
 		MainMemory memory = new MainMemory(16);
 		String[] instructions = { "ADD R2, R4, R5", "SUB R9, R8, R7", "MUL R2, R4, R4", "DIV R0, R1, R5", "ADDI R2, R2, 45" };
-		
+
 		memory.readProgram(instructions, 0);
 		for (int i = 0; i < instructions.length; i++) {
 			System.out.println(memory.readBlock(i * 4));
 		}
 	}
+
 }
